+++ lib/src/interactive/scene_controller.dart
@@ -1,6 +1,8 @@
+import 'dart:ui';
+
 import 'package:flutter/foundation.dart';
 
-import '../contract/scene_render_state.dart';
+import '../contract/scene_view_render_state.dart';
 import '../contract/snapshot.dart';
 import '../core/action_events.dart';
 import '../core/pointer_input.dart';
@@ -12,7 +14,7 @@ import 'scene_controller_interaction.dart';
 import 'scene_controller_scene.dart';
 import 'scene_controller_selection.dart';
 
-class SceneController extends ChangeNotifier implements SceneRenderState {
+class SceneController extends ChangeNotifier implements SceneViewRenderState {
   SceneController({
     SceneSnapshot? initialSnapshot,
     PointerInputSettings? pointerSettings,
@@ -67,6 +69,53 @@ class SceneController extends ChangeNotifier implements SceneRenderState {
 
   @override
   Set<NodeId> get selectedNodeIds => _core.selectedNodeIds;
+
+  @override
+  int get controllerEpoch => _core.controllerEpoch;
+
+  @override
+  Rect? get selectionRect => _facade.interactionRuntime.selectionRect;
+
+  @override
+  Offset get cameraOffset => snapshot.camera.offset;
+
+  @override
+  Offset Function(NodeId nodeId) get previewDeltaResolver =>
+      _facade.interactionRuntime.previewDeltaForNode;
+
+  @override
+  bool get hasActiveStrokePreview => interaction.hasActiveStrokePreview;
+
+  @override
+  List<Offset> get activeStrokePreviewPoints =>
+      interaction.activeStrokePreviewPoints;
+
+  @override
+  double get activeStrokePreviewThickness =>
+      interaction.activeStrokePreviewThickness;
+
+  @override
+  Color get activeStrokePreviewColor => interaction.activeStrokePreviewColor;
+
+  @override
+  double get activeStrokePreviewOpacity =>
+      interaction.activeStrokePreviewOpacity;
+
+  @override
+  bool get hasActiveLinePreview => interaction.hasActiveLinePreview;
+
+  @override
+  Offset? get activeLinePreviewStart => interaction.activeLinePreviewStart;
+
+  @override
+  Offset? get activeLinePreviewEnd => interaction.activeLinePreviewEnd;
+
+  @override
+  double get activeLinePreviewThickness =>
+      interaction.activeLinePreviewThickness;
+
+  @override
+  Color get activeLinePreviewColor => interaction.activeLinePreviewColor;
 
   SceneControllerInteraction get interaction => _facade.interaction;
   SceneControllerSelection get selection => _facade.selection;